+++ web/CHANGELOG.md
@@ -11,6 +11,8 @@
 - Added `childNodesAsList` to `Node` and `childrenAsList` to `Element` via 
   extensions to support mutable operations on node lists.
 - Added `asList` to `NodeList` via extension.
+- Added `data` to `HTMLElement` and `SVGElement` via extension. It provides 
+  nullable interface for `dataset` allowing to check if data attribute is set.
 
 ## 1.1.1